backend/requirements.txt
minio
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2